+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Release History
 
 ## [0.4.0] - unreleased
+- [#62](https://github.com/kwila-cloud/simple-sync/pull/62): Add storage documentation
+- [#61](https://github.com/kwila-cloud/simple-sync/pull/61): Add performance and concurrency tests for SQLite storage
 - [#60](https://github.com/kwila-cloud/simple-sync/pull/60): Implement SQLite setup token and API key storage
 - [#59](https://github.com/kwila-cloud/simple-sync/pull/59): Implement SQLite ACL rule storage
 - [#58](https://github.com/kwila-cloud/simple-sync/pull/58): Implement SQLite user storage